+++ huffman/src/decode/table2.rs
@@ -1,7 +1,7 @@
 /// This is a static translation table for decoding Huffman sequence by reading
 /// 2-bit(s) at at time.
 #[cfg(feature = "decode2")]
-pub const DECODE_TABLE: [[(Option<usize>, Option<usize>, usize); 4]; 126] = [ // (next_id, ascii, leftover)
+pub const DECODE_TABLE: [[(Option<u8>, Option<u16>, u8); 4]; 126] = [ // (next_id, ascii, leftover)
     [ // 0
         (Some(36), None, 0),
         (Some(29), None, 0),
