+++ huffman/src/decode/table4.rs
@@ -1,7 +1,7 @@
 /// This is a static translation table for decoding Huffman sequence by reading
 /// 4-bit(s) at at time.
 #[cfg(feature = "decode4")]
-pub const DECODE_TABLE: [[(Option<usize>, Option<usize>, usize); 16]; 54] = [ // (next_id, ascii, leftover)
+pub const DECODE_TABLE: [[(Option<u8>, Option<u16>, u8); 16]; 54] = [ // (next_id, ascii, leftover)
     [ // 0
         (Some(15), None, 0),
         (Some(16), None, 0),
